requirements.txt
agentops==0.3.21
anthropic==0.49.0
apify_client==1.9.2
arxiv==2.1.3
arxiv2text==0.1.14
azure_storage==0.37.0
beautifulsoup4==4.13.3
botocore==1.35.79
chroma_hnswlib==0.7.6
chunkr_ai==0.0.41
cohere==5.14.0
colorama==0.4.6
datacommons==1.4.3
datacommons_pandas==0.0.3
datasets==2.21.0
discord.py==2.4.0
docker_py==1.10.6
docstring_parser==0.16
docx2markdown==0.1.1
docx2txt==0.8
duckduckgo_search==7.2.1
fastapi==0.115.11
ffmpeg_python==0.2.0
firecrawl_py==1.11.0
fish_audio_sdk==2025.2.11
fitz==0.0.1.dev2
googlemaps==4.10.0
gradio==5.20.0
graspologic==3.4.1
hnswlib==0.8.0
html2text==2024.2.26
httpx==0.28.1
huggingface_hub==0.26.5
imageio==2.35.0
jsonschema==4.23.0
jupyter_client==8.6.3
langchain==0.3.19
langchain_community==0.3.18
librosa==0.10.2.post1
litellm==1.62.1
loguru==0.7.3
mistral_common==1.5.3
mistralai==1.5.0
mock==5.2.0
neo4j==5.27.0
nest_asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
notion_client==2.2.1
numpy==2.2.3
opencv_python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
Pillow==11.1.0
prance==23.6.21.0
praw==7.8.1
protobuf==5.29.3
pydantic==2.10.6
pydub==0.25.1
PyGithub==2.5.0
PyGithub==2.6.1
pymilvus==2.5.0
pyowm==3.3.0
PyPDF2==3.0.1
pyTelegramBotAPI==4.26.0
pytest==8.3.4
python_pptx==0.6.23
PyYAML==6.0.2
PyYAML==6.0.2
qdrant_client==1.13.2
qwen_vl_utils==0.0.10
rank_bm25==0.2.2
Requests==2.32.3
requests_oauthlib==1.3.1
retry==0.9.2
scenedetect==0.6.5.2
scholarly==1.7.11
sentence_transformers==3.3.1
slack_bolt==1.21.3
slack_sdk==3.33.5
SQLAlchemy==2.0.36
starlette==0.46.0
streamlit==1.42.2
streamlit_chat==0.1.1
sympy==1.13.1
tabulate==0.9.0
tavily_python==0.5.0
text2vec==1.3.2
textblob==0.19.0
tiktoken==0.7.0
torch==2.2.0
tqdm==4.67.1
transformers==4.49.0.dev0
unstructured==0.14.10
uvicorn==0.34.0
wikipedia==1.4.0
wolframalpha==5.1.3
xls2xlsx==0.2.0
xmltodict==0.14.2
xxhash==3.5.0
yt_dlp==2025.2.19
